requirements.txt
pandas>=1.1
matplotlib>=3.0
dnspython>=2.0
